Manage contacts by owner

+++ helpers/helpers.js
@@ -1,12 +1,12 @@
 import { Contact } from '../models/schemas/contact.js';
 import { User } from '../models/schemas/user.js';
 
-const getAllContacts = () => Contact.find();
-const findAndDeleteContact = id => Contact.findByIdAndDelete(id);
+const getAllContacts = user => Contact.find({ owner: user._id });
+const findAndDeleteContact = (id, user) => Contact.findByIdAndDelete({ id, owner: user._id });
 const addNewContact = contact => Contact.create(contact);
-const findContact = filter => Contact.findOne(filter);
-const findAndUpdateContact = (filter, updateField) =>
-  Contact.findOneAndUpdate(filter, updateField, { new: true });
+const findContact = (id, user) => Contact.findOne({ _id: id, owner: user._id });
+const findAndUpdateContact = (id, user, updateField) =>
+  Contact.findOneAndUpdate({ _id: id, owner: user._id }, updateField, { new: true });
 
 const findUser = filter => User.findOne(filter);
 const findAndUpdateUser = (filter, updateField) =>
